refactor(interpreter): break down interpreter
